perf(gguf): let dequant follow the model dtype, and measure the Turing bf16 penalty
GGUFEmbedding hardcoded bf16 for its dequant output, which silently forced the whole
model back to bf16 whatever --dtype asked for. It now takes the compute dtype explicitly,
read from a tensor the engine allocated under its own torch_dtype(config.dtype) context
rather than from the ambient default, which is only the model dtype while that context is
open. The qwen4exp hyper-connection mixer likewise computes in the weights' dtype instead
of letting grouped_rms_norm's float32 accumulation reach the GEMM.

This matters more than it looks on Turing. sm_75 has no bf16 tensor cores, and a
microbenchmark on the Quadro RTX 6000 shows batched bf16 GEMMs falling off a cliff:

  shape                      bf16       fp16     fp16 speedup
  decode bs=1              34.8 us    36.0 us      0.97x
  bs=4                    864.2 us    39.5 us     21.88x
  prefill 300             791.1 us    64.3 us     12.31x
  square 2560, prefill    526.7 us    87.5 us      6.02x

which also explains why concurrency 4 only bought 1.19x aggregate.

fp16 end-to-end is NOT enabled yet, and this commit does not claim it. With these fixes
the failure moves past the embedding and the mixer into the GDN, whose fla recurrence
kernels emit float32 against an fp16 out_proj. Making that path dtype-generic is shared
surface with Ornith and carries real numerical risk, so it is scoped work rather than
something to rush. The changes here are correct in bf16 regardless and are the
prerequisite for it.

Also correcting an earlier claim of mine: I reported the chat endpoint as returning empty
content. It does not -- content is 'Paris' with the thinking split into reasoning_content.
My test used max_tokens=48, which truncated the model inside its <think> block. That said,
with an adequate budget most prompts still end the turn after reasoning without emitting
content, and the stop tokens are verified correct (EOS 248046 = <|im_end|>), so there is a
real chat-template/thinking-mode issue underneath -- just not the one I described.

469 tests pass; the 2 FP8 failures are pre-existing and unrelated.

+++ python/freetoken/models/qwen4exp/hyper_connections.py
@@ -71,14 +71,17 @@ def hc_mix(
     ``w_inject`` is None for the final mix before the LM head, which produces no injection.
     """
     T, _, D = state.shape
-    xn = grouped_rms_norm(state, w_norm, eps, hc)          # [T, hc*D]
+    # Compute in the weights' dtype. grouped_rms_norm accumulates in at least float32, and
+    # letting that choice reach the GEMM both mismatches the weights and would silently
+    # pick a slower kernel -- on Turing a batched bf16 GEMM is 12-22x slower than fp16.
+    xn = grouped_rms_norm(state, w_norm, eps, hc).to(w_down.dtype)   # [T, hc*D]
 
     # The 1/hc is applied before the SiLU, so it is not a rescale of the gate.
     lo = F.silu(F.linear(xn, w_down) / hc)                 # [T, low_rank]
     gate = torch.sigmoid(F.linear(lo, w_up))               # [T, hc*D]
 
     gated = (xn * gate).reshape(T, hc, D)
-    mixed = gated.mean(dim=1)                              # [T, D]
+    mixed = gated.mean(dim=1).to(state.dtype)              # [T, D]
 
     inject = F.linear(xn, w_inject) if w_inject is not None else None
     return mixed, inject
